Abstract and add to flow controller.
